Rename test file to align with moved functions
